backend/requirements.txt
fastapi>=0.124.1,<1.0.0
uvicorn>=0.34.0,<1.0.0
google-adk==1.25.1
langgraph==1.0.9
google-cloud-aiplatform[agent-engines]
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
google-cloud-discoveryengine>=0.11.0
langchain-google-community>=1.0.0
python-dotenv==1.0.1
httpx==0.28.1
pillow==10.4.0
python-multipart==0.0.12
slowapi==0.1.9
cachetools==5.3.3
tenacity>=9.0.0,<10.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
certifi==2024.2.2
deep-translator>=1.11.0
websockets>=12.0
requests>=2.31.0
